+++ Source/ACE.Database/ACE.Database.csproj
@@ -73,13 +73,13 @@
     <Compile Include="ICommonDatabase.cs" />
     <Compile Include="IShardDatabase.cs" />
     <Compile Include="IWorldDatabase.cs" />
-    <Compile Include="LightweightEntities\SlimCharacter.cs" />
     <Compile Include="MySqlResult.cs">
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ShardDatabase.cs" />
     <Compile Include="StoredPreparedStatement.cs" />
-    <Compile Include="CommonDatabase.cs" />
+    <Compile Include="WorldDatabase.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
@@ -94,9 +94,6 @@
       <Project>{aa611a84-cdf8-4ab6-8d53-085c897a7422}</Project>
       <Name>ACE.Entity</Name>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="LightweightEntities\Readme.txt" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\StyleCop.MSBuild.5.0.0-beta01\build\StyleCop.MSBuild.targets" Condition="Exists('..\packages\StyleCop.MSBuild.5.0.0-beta01\build\StyleCop.MSBuild.targets')" />
