+++ Source/ACE.Database/IWorldDatabase.cs
@@ -14,6 +14,11 @@ namespace ACE.Database
         List<AceCreatureStaticLocation> GetCreaturesByLandblock(ushort landblock);
 
         List<AceCreatureGeneratorLocation> GetCreatureGeneratorsByLandblock(ushort landblock);
+        
+        bool InsertStaticCreatureLocation(AceCreatureStaticLocation acsl);
 
+        AceObject GetRandomWeenieOfType(uint typeId);
+
+        AceObject GetWeenie(uint weenieClassId);
     }
 }
