--- Source/ACE.Database/LightweightEntities/Readme.txt
@@ -1,4 +1,0 @@
-﻿Lightweight Entities are:
-* Only used for internal ORM operations
-* Always internal to ACE.Database
-* Named "Slim[Entity]"
