+++ Source/ACE/Entity/Creature.cs
@@ -51,12 +51,12 @@ namespace ACE.Entity
         }
 
         public Creature(AceCreatureStaticLocation aceC)
-            : base((ObjectType)aceC.CreatureData.ItemType,
+            : base((ObjectType)aceC.WeenieObject.ItemType,
                   new ObjectGuid(CommonObjectFactory.DynamicObjectId, GuidType.Creature),
-                  aceC.CreatureData.Name,
+                  aceC.WeenieObject.Name,
                   aceC.WeenieClassId,
-                  (ObjectDescriptionFlag)aceC.CreatureData.AceObjectDescriptionFlags,
-                  (WeenieHeaderFlag)aceC.CreatureData.WeenieHeaderFlags,
+                  (ObjectDescriptionFlag)aceC.WeenieObject.AceObjectDescriptionFlags,
+                  (WeenieHeaderFlag)aceC.WeenieObject.WeenieHeaderFlags,
                   aceC.Position)
         {
             if (aceC.WeenieClassId < 0x8000u)
@@ -64,11 +64,11 @@ namespace ACE.Entity
             else
                 this.WeenieClassid = (ushort)(aceC.WeenieClassId - 0x8000);
 
-            SetObjectData(aceC.CreatureData);
-            SetAbilities(aceC.CreatureData);
+            SetObjectData(aceC.WeenieObject);
+            SetAbilities(aceC.WeenieObject);
         }
 
-        private void SetObjectData(AceCreatureObject aco)
+        private void SetObjectData(AceObject aco)
         {
             PhysicsData.CurrentMotionState = new UniversalMotion(MotionStance.Standing);
             PhysicsData.MTableResourceId = aco.MotionTableId;
@@ -96,7 +96,7 @@ namespace ACE.Entity
             // ModelData.PaletteGuid = aco.PaletteId;
         }
 
-        private void SetAbilities(AceCreatureObject aco)
+        private void SetAbilities(AceObject aco)
         {
             Strength = new CreatureAbility(aco, Enum.Ability.Strength);
             Endurance = new CreatureAbility(aco, Enum.Ability.Endurance);
