+++ Source/ACE/Entity/Player.cs
@@ -132,11 +132,6 @@ namespace ACE.Entity
             set { Character.IsPsr = value; }
         }
         
-        //public ReadOnlyDictionary<Skill, CreatureSkill> Skills
-        //{
-        //    get { return new ReadOnlyDictionary<Skill, CreatureSkill>(Character.Skills); }
-        //}
-
         public uint TotalLogins
         {
             get { return Character.TotalLogins; }
@@ -212,266 +207,214 @@ namespace ACE.Entity
         {
             // TODO: this needs a new server message to decode the data and THEN create a new AceCharacter object
 
-            //            Character character = new Character();
-
-            //            reader.Skip(4);   /* Unknown constant (1) */
-
-            //            character.Heritage = reader.ReadUInt32();
-            //            character.Gender = reader.ReadUInt32();
-            //            character.Appearance = Appearance.FromNetowrk(reader);
-            //            character.TemplateOption = reader.ReadUInt32();
-            //            character.StrengthAbility.Base = reader.ReadUInt32();
-            //            character.EnduranceAbility.Base = reader.ReadUInt32();
-            //            character.CoordinationAbility.Base = reader.ReadUInt32();
-            //            character.QuicknessAbility.Base = reader.ReadUInt32();
-            //            character.FocusAbility.Base = reader.ReadUInt32();
-            //            character.SelfAbility.Base = reader.ReadUInt32();
-            //            character.Slot = reader.ReadUInt32();
-            //            character.ClassId = reader.ReadUInt32();
-
-            //            // characters start with max vitals
-            //            character.Health.Current = character.Health.UnbuffedValue;
-            //            character.Stamina.Current = character.Stamina.UnbuffedValue;
-            //            character.Mana.Current = character.Mana.UnbuffedValue;
-
-            //            character.TotalSkillCredits = 52;
-            //            character.AvailableSkillCredits = 52;
-
-            //            uint numOfSkills = reader.ReadUInt32();
-            //            Skill skill;
-            //            SkillStatus skillStatus;
-            //            SkillCostAttribute skillCost;
-            //            for (uint i = 0; i < numOfSkills; i++)
-            //            {
-            //                skill = (Skill)i;
-            //                skillCost = skill.GetCost();
-            //                skillStatus = (SkillStatus)reader.ReadUInt32();
-            //                if (skillStatus == SkillStatus.Specialized)
-            //                    if (skillCost.TrainsFree)
-            //                        character.AvailableSkillCredits = character.AvailableSkillCredits - skillCost.SpecializationCost;
-            //                    else
-            //                        character.AvailableSkillCredits = character.AvailableSkillCredits - skillCost.SpecializationCost - skillCost.TrainingCost;
-            //                if (skillStatus == SkillStatus.Trained)
-            //                    if (!skillCost.TrainsFree)
-            //                        character.AvailableSkillCredits = character.AvailableSkillCredits - skillCost.TrainingCost;
-            //                character.AddSkill(skill, skillStatus, 0, 0);
-            //            }
-
-            //            character.Name = reader.ReadString16L();
-            //            character.StartArea = reader.ReadUInt32();
-            //            character.IsAdmin = Convert.ToBoolean(reader.ReadUInt32());
-            //            character.IsEnvoy = Convert.ToBoolean(reader.ReadUInt32());
-            //            // character.TotalSkillPoints = reader.ReadUInt32();  // garbage ? 
-
-            //            return character;
-
-
             // TODO: Move this all into Character Creation and store directly in the database.
-            //if (DatManager.PortalDat.AllFiles.ContainsKey(0x0E000002))
-            //{
-            //    CharGen cg = CharGen.ReadFromDat();
 
-            //    int h = Convert.ToInt32(character.PropertiesInt[PropertyInt.HeritageGroup]);
-            //    int s = Convert.ToInt32(character.PropertiesInt[PropertyInt.Gender]);
-            //    SexCG sex = cg.HeritageGroups[h].SexList[s];
-            //    // Set the character basics
-            //    PhysicsData.MTableResourceId = sex.MotionTable;
-            //    PhysicsData.Stable = sex.SoundTable;
-            //    PhysicsData.Petable = sex.PhysicsTable;
-            //    PhysicsData.CSetup = sex.SetupID;
-            //    ModelData.PaletteGuid = sex.BasePalette;
+            /* keeping this block in case Optim needs it
+            if (DatManager.PortalDat.AllFiles.ContainsKey(0x0E000002))
+            {
+                CharGen cg = CharGen.ReadFromDat();
 
-            //    // Check the character scale
-            //    if (sex.Scale != 100u)
-            //    {
-            //        // Set the PhysicsData flag to let it know we're changing the scale
-            //        PhysicsData.PhysicsDescriptionFlag |= PhysicsDescriptionFlag.ObjScale;
-            //        PhysicsData.ObjScale = sex.Scale / 100f; // Scale is stored as a percentage
-            //    }
+                int h = Convert.ToInt32(character.PropertiesInt[PropertyInt.HeritageGroup]);
+                int s = Convert.ToInt32(character.PropertiesInt[PropertyInt.Gender]);
+                SexCG sex = cg.HeritageGroups[h].SexList[s];
+                // Set the character basics
+                PhysicsData.MTableResourceId = sex.MotionTable;
+                PhysicsData.Stable = sex.SoundTable;
+                PhysicsData.Petable = sex.PhysicsTable;
+                PhysicsData.CSetup = sex.SetupID;
+                ModelData.PaletteGuid = sex.BasePalette;
 
-            //    // Get the hair first, because we need to know if you're bald, and that's the name of that tune!
-            //    HairStyleCG hairstyle = sex.HairStyleList[Convert.ToInt32(character.Appearance.HairStyle)];
-            //    bool isBald = hairstyle.Bald;
+                // Check the character scale
+                if (sex.Scale != 100u)
+                {
+                    // Set the PhysicsData flag to let it know we're changing the scale
+                    PhysicsData.PhysicsDescriptionFlag |= PhysicsDescriptionFlag.ObjScale;
+                    PhysicsData.ObjScale = sex.Scale / 100f; // Scale is stored as a percentage
+                }
 
-            //    // Certain races (Undead, Tumeroks, Others?) have multiple body styles available. This is controlled via the "hair style".
-            //    if (hairstyle.AlternateSetup > 0)
-            //        PhysicsData.CSetup = hairstyle.AlternateSetup;
+                // Get the hair first, because we need to know if you're bald, and that's the name of that tune!
+                HairStyleCG hairstyle = sex.HairStyleList[Convert.ToInt32(character.Appearance.HairStyle)];
+                bool isBald = hairstyle.Bald;
 
-            //    // Apply the hair models & texture changes
-            //    for (int i = 0; i < hairstyle.ObjDesc.AnimPartChanges.Count; i++)
-            //        ModelData.AddModel(hairstyle.ObjDesc.AnimPartChanges[i].PartIndex, hairstyle.ObjDesc.AnimPartChanges[i].PartID);
-            //    for (int i = 0; i < hairstyle.ObjDesc.TextureChanges.Count; i++)
-            //        ModelData.AddTexture(hairstyle.ObjDesc.TextureChanges[i].PartIndex, hairstyle.ObjDesc.TextureChanges[i].OldTexture, hairstyle.ObjDesc.TextureChanges[i].NewTexture);
+                // Certain races (Undead, Tumeroks, Others?) have multiple body styles available. This is controlled via the "hair style".
+                if (hairstyle.AlternateSetup > 0)
+                    PhysicsData.CSetup = hairstyle.AlternateSetup;
 
-            //    // Eyes only have Texture Changes - eye color is set seperately
-            //    ObjDesc eyes;
-            //    if (hairstyle.Bald)
-            //        eyes = sex.EyeStripList[Convert.ToInt32(character.Appearance.Eyes)].ObjDescBald;
-            //    else
-            //        eyes = sex.EyeStripList[Convert.ToInt32(character.Appearance.Eyes)].ObjDesc;
-            //    for (int i = 0; i < eyes.TextureChanges.Count; i++)
-            //        ModelData.AddTexture(eyes.TextureChanges[i].PartIndex, eyes.TextureChanges[i].OldTexture, eyes.TextureChanges[i].NewTexture);
+                // Apply the hair models & texture changes
+                for (int i = 0; i < hairstyle.ObjDesc.AnimPartChanges.Count; i++)
+                    ModelData.AddModel(hairstyle.ObjDesc.AnimPartChanges[i].PartIndex, hairstyle.ObjDesc.AnimPartChanges[i].PartID);
+                for (int i = 0; i < hairstyle.ObjDesc.TextureChanges.Count; i++)
+                    ModelData.AddTexture(hairstyle.ObjDesc.TextureChanges[i].PartIndex, hairstyle.ObjDesc.TextureChanges[i].OldTexture, hairstyle.ObjDesc.TextureChanges[i].NewTexture);
 
-            //    // Nose only has Texture Changes
-            //    ObjDesc nose = sex.NoseStripList[Convert.ToInt32(character.Appearance.Nose)].ObjDesc;
-            //    for (int i = 0; i < nose.TextureChanges.Count; i++)
-            //        ModelData.AddTexture(nose.TextureChanges[i].PartIndex, nose.TextureChanges[i].OldTexture, nose.TextureChanges[i].NewTexture);
+                // Eyes only have Texture Changes - eye color is set seperately
+                ObjDesc eyes;
+                if (hairstyle.Bald)
+                    eyes = sex.EyeStripList[Convert.ToInt32(character.Appearance.Eyes)].ObjDescBald;
+                else
+                    eyes = sex.EyeStripList[Convert.ToInt32(character.Appearance.Eyes)].ObjDesc;
+                for (int i = 0; i < eyes.TextureChanges.Count; i++)
+                    ModelData.AddTexture(eyes.TextureChanges[i].PartIndex, eyes.TextureChanges[i].OldTexture, eyes.TextureChanges[i].NewTexture);
 
-            //    // Mouth, suprise, only Texture Changes
-            //    ObjDesc mouth = sex.MouthStripList[Convert.ToInt32(character.Appearance.Mouth)].ObjDesc;
-            //    for (int i = 0; i < mouth.TextureChanges.Count; i++)
-            //        ModelData.AddTexture(mouth.TextureChanges[i].PartIndex, mouth.TextureChanges[i].OldTexture, mouth.TextureChanges[i].NewTexture);
+                // Nose only has Texture Changes
+                ObjDesc nose = sex.NoseStripList[Convert.ToInt32(character.Appearance.Nose)].ObjDesc;
+                for (int i = 0; i < nose.TextureChanges.Count; i++)
+                    ModelData.AddTexture(nose.TextureChanges[i].PartIndex, nose.TextureChanges[i].OldTexture, nose.TextureChanges[i].NewTexture);
 
-            //    // Skin is stored as PaletteSet (list of Palettes), so we need to read in the set to get the specific palette
-            //    PaletteSet skinPalSet = PaletteSet.ReadFromDat(sex.SkinPalSet);
-            //    ushort skinPal = (ushort)skinPalSet.GetPaletteID(character.Appearance.SkinHue);
-            //    // Apply the skin palette...
-            //    ModelData.AddPalette(skinPal, 0x0, 0x18);
+                // Mouth, suprise, only Texture Changes
+                ObjDesc mouth = sex.MouthStripList[Convert.ToInt32(character.Appearance.Mouth)].ObjDesc;
+                for (int i = 0; i < mouth.TextureChanges.Count; i++)
+                    ModelData.AddTexture(mouth.TextureChanges[i].PartIndex, mouth.TextureChanges[i].OldTexture, mouth.TextureChanges[i].NewTexture);
 
-            //    // Hair is stored as PaletteSet (list of Palettes), so we need to read in the set to get the specific palette
-            //    PaletteSet hairPalSet = PaletteSet.ReadFromDat(sex.HairColorList[Convert.ToInt32(character.Appearance.HairColor)]);
-            //    ushort hairPal = (ushort)hairPalSet.GetPaletteID(character.Appearance.HairHue);
-            //    ModelData.AddPalette(hairPal, 0x18, 0x8);
+                // Skin is stored as PaletteSet (list of Palettes), so we need to read in the set to get the specific palette
+                PaletteSet skinPalSet = PaletteSet.ReadFromDat(sex.SkinPalSet);
+                ushort skinPal = (ushort)skinPalSet.GetPaletteID(character.Appearance.SkinHue);
+                // Apply the skin palette...
+                ModelData.AddPalette(skinPal, 0x0, 0x18);
 
-            //    // Eye color palette
-            //    ModelData.AddPalette(sex.EyeColorList[Convert.ToInt32(character.Appearance.EyeColor)], 0x20, 0x8);
+                // Hair is stored as PaletteSet (list of Palettes), so we need to read in the set to get the specific palette
+                PaletteSet hairPalSet = PaletteSet.ReadFromDat(sex.HairColorList[Convert.ToInt32(character.Appearance.HairColor)]);
+                ushort hairPal = (ushort)hairPalSet.GetPaletteID(character.Appearance.HairHue);
+                ModelData.AddPalette(hairPal, 0x18, 0x8);
 
-            //    // Get the character's startup gear.
-            //    // TODO: Load the proper inventory/equipment options once that system is created.
-            //    if (character.Appearance.HeadgearStyle < 0xFFFFFFFF) // No headgear is max UINT
-            //    {
-            //        uint headgearTableID = sex.HeadgearList[Convert.ToInt32(character.Appearance.HeadgearStyle)].ClothingTable;
-            //        ClothingTable headCT = ClothingTable.ReadFromDat(headgearTableID);
-            //        if (headCT.ClothingBaseEffects.ContainsKey(sex.SetupID))
-            //        {
-            //            // Add the model and texture(s)
-            //            ClothingBaseEffect headCBE = headCT.ClothingBaseEffects[sex.SetupID];
-            //            for (int i = 0; i < headCBE.CloObjectEffects.Count; i++)
-            //            {
-            //                byte partNum = (byte)headCBE.CloObjectEffects[i].Index;
-            //                ModelData.AddModel((byte)headCBE.CloObjectEffects[i].Index, (ushort)headCBE.CloObjectEffects[i].ModelId);
+                // Eye color palette
+                ModelData.AddPalette(sex.EyeColorList[Convert.ToInt32(character.Appearance.EyeColor)], 0x20, 0x8);
 
-            //                for (int j = 0; j < headCBE.CloObjectEffects[i].CloTextureEffects.Count; j++)
-            //                    ModelData.AddTexture((byte)headCBE.CloObjectEffects[i].Index, (ushort)headCBE.CloObjectEffects[i].CloTextureEffects[j].OldTexture, (ushort)headCBE.CloObjectEffects[i].CloTextureEffects[j].NewTexture);
-            //            }
+                // Get the character's startup gear.
+                // TODO: Load the proper inventory/equipment options once that system is created.
+                if (character.Appearance.HeadgearStyle < 0xFFFFFFFF) // No headgear is max UINT
+                {
+                    uint headgearTableID = sex.HeadgearList[Convert.ToInt32(character.Appearance.HeadgearStyle)].ClothingTable;
+                    ClothingTable headCT = ClothingTable.ReadFromDat(headgearTableID);
+                    if (headCT.ClothingBaseEffects.ContainsKey(sex.SetupID))
+                    {
+                        // Add the model and texture(s)
+                        ClothingBaseEffect headCBE = headCT.ClothingBaseEffects[sex.SetupID];
+                        for (int i = 0; i < headCBE.CloObjectEffects.Count; i++)
+                        {
+                            byte partNum = (byte)headCBE.CloObjectEffects[i].Index;
+                            ModelData.AddModel((byte)headCBE.CloObjectEffects[i].Index, (ushort)headCBE.CloObjectEffects[i].ModelId);
 
-            //            // Apply the proper palette(s). Unlike character skin/hair, clothes can have several palette ranges!
-            //            CloSubPalEffect headSubPal = headCT.ClothingSubPalEffects[character.Appearance.HeadgearColor];
-            //            for (int i = 0; i < headSubPal.CloSubPalettes.Count; i++)
-            //            {
-            //                PaletteSet headgearPalSet = PaletteSet.ReadFromDat(headSubPal.CloSubPalettes[i].PaletteSet);
-            //                ushort headgearPal = (ushort)headgearPalSet.GetPaletteID(character.Appearance.HeadgearHue);
+                            for (int j = 0; j < headCBE.CloObjectEffects[i].CloTextureEffects.Count; j++)
+                                ModelData.AddTexture((byte)headCBE.CloObjectEffects[i].Index, (ushort)headCBE.CloObjectEffects[i].CloTextureEffects[j].OldTexture, (ushort)headCBE.CloObjectEffects[i].CloTextureEffects[j].NewTexture);
+                        }
 
-            //                for (int j = 0; j < headSubPal.CloSubPalettes[i].Ranges.Count; j++)
-            //                {
-            //                    uint palOffset = headSubPal.CloSubPalettes[i].Ranges[j].Offset / 8;
-            //                    uint numColors = headSubPal.CloSubPalettes[i].Ranges[j].NumColors / 8;
-            //                    ModelData.AddPalette(headgearPal, (ushort)palOffset, (ushort)numColors);
-            //                }
-            //            }
-            //        }
-            //    }
+                        // Apply the proper palette(s). Unlike character skin/hair, clothes can have several palette ranges!
+                        CloSubPalEffect headSubPal = headCT.ClothingSubPalEffects[character.Appearance.HeadgearColor];
+                        for (int i = 0; i < headSubPal.CloSubPalettes.Count; i++)
+                        {
+                            PaletteSet headgearPalSet = PaletteSet.ReadFromDat(headSubPal.CloSubPalettes[i].PaletteSet);
+                            ushort headgearPal = (ushort)headgearPalSet.GetPaletteID(character.Appearance.HeadgearHue);
 
-            //    // Get the character's initial pants
-            //    uint pantsTableID = sex.PantsList[Convert.ToInt32(character.Appearance.PantsStyle)].ClothingTable;
-            //    ClothingTable pantsCT = ClothingTable.ReadFromDat(pantsTableID);
-            //    if (pantsCT.ClothingBaseEffects.ContainsKey(sex.SetupID))
-            //    {
-            //        ClothingBaseEffect pantsCBE = pantsCT.ClothingBaseEffects[sex.SetupID];
-            //        for (int i = 0; i < pantsCBE.CloObjectEffects.Count; i++)
-            //        {
-            //            byte partNum = (byte)pantsCBE.CloObjectEffects[i].Index;
-            //            ModelData.AddModel((byte)pantsCBE.CloObjectEffects[i].Index, (ushort)pantsCBE.CloObjectEffects[i].ModelId);
+                            for (int j = 0; j < headSubPal.CloSubPalettes[i].Ranges.Count; j++)
+                            {
+                                uint palOffset = headSubPal.CloSubPalettes[i].Ranges[j].Offset / 8;
+                                uint numColors = headSubPal.CloSubPalettes[i].Ranges[j].NumColors / 8;
+                                ModelData.AddPalette(headgearPal, (ushort)palOffset, (ushort)numColors);
+                            }
+                        }
+                    }
+                }
 
-            //            for (int j = 0; j < pantsCBE.CloObjectEffects[i].CloTextureEffects.Count; j++)
-            //                ModelData.AddTexture((byte)pantsCBE.CloObjectEffects[i].Index, (ushort)pantsCBE.CloObjectEffects[i].CloTextureEffects[j].OldTexture, (ushort)pantsCBE.CloObjectEffects[i].CloTextureEffects[j].NewTexture);
-            //        }
+                // Get the character's initial pants
+                uint pantsTableID = sex.PantsList[Convert.ToInt32(character.Appearance.PantsStyle)].ClothingTable;
+                ClothingTable pantsCT = ClothingTable.ReadFromDat(pantsTableID);
+                if (pantsCT.ClothingBaseEffects.ContainsKey(sex.SetupID))
+                {
+                    ClothingBaseEffect pantsCBE = pantsCT.ClothingBaseEffects[sex.SetupID];
+                    for (int i = 0; i < pantsCBE.CloObjectEffects.Count; i++)
+                    {
+                        byte partNum = (byte)pantsCBE.CloObjectEffects[i].Index;
+                        ModelData.AddModel((byte)pantsCBE.CloObjectEffects[i].Index, (ushort)pantsCBE.CloObjectEffects[i].ModelId);
 
-            //        // Apply the proper palette(s). Unlike character skin/hair, clothes can have several palette ranges!
-            //        CloSubPalEffect pantsSubPal = pantsCT.ClothingSubPalEffects[character.Appearance.PantsColor];
-            //        for (int i = 0; i < pantsSubPal.CloSubPalettes.Count; i++)
-            //        {
-            //            PaletteSet pantsPalSet = PaletteSet.ReadFromDat(pantsSubPal.CloSubPalettes[i].PaletteSet);
-            //            ushort pantsPal = (ushort)pantsPalSet.GetPaletteID(character.Appearance.PantsHue);
+                        for (int j = 0; j < pantsCBE.CloObjectEffects[i].CloTextureEffects.Count; j++)
+                            ModelData.AddTexture((byte)pantsCBE.CloObjectEffects[i].Index, (ushort)pantsCBE.CloObjectEffects[i].CloTextureEffects[j].OldTexture, (ushort)pantsCBE.CloObjectEffects[i].CloTextureEffects[j].NewTexture);
+                    }
 
-            //            for (int j = 0; j < pantsSubPal.CloSubPalettes[i].Ranges.Count; j++)
-            //            {
-            //                uint palOffset = pantsSubPal.CloSubPalettes[i].Ranges[j].Offset / 8;
-            //                uint numColors = pantsSubPal.CloSubPalettes[i].Ranges[j].NumColors / 8;
-            //                ModelData.AddPalette(pantsPal, (ushort)palOffset, (ushort)numColors);
-            //            }
-            //        }
-            //    } // end pants
+                    // Apply the proper palette(s). Unlike character skin/hair, clothes can have several palette ranges!
+                    CloSubPalEffect pantsSubPal = pantsCT.ClothingSubPalEffects[character.Appearance.PantsColor];
+                    for (int i = 0; i < pantsSubPal.CloSubPalettes.Count; i++)
+                    {
+                        PaletteSet pantsPalSet = PaletteSet.ReadFromDat(pantsSubPal.CloSubPalettes[i].PaletteSet);
+                        ushort pantsPal = (ushort)pantsPalSet.GetPaletteID(character.Appearance.PantsHue);
 
-            //    // Get the character's initial shirt
-            //    uint shirtTableID = sex.ShirtList[Convert.ToInt32(character.Appearance.ShirtStyle)].ClothingTable;
-            //    ClothingTable shirtCT = ClothingTable.ReadFromDat(shirtTableID);
-            //    if (shirtCT.ClothingBaseEffects.ContainsKey(sex.SetupID))
-            //    {
-            //        ClothingBaseEffect shirtCBE = shirtCT.ClothingBaseEffects[sex.SetupID];
-            //        for (int i = 0; i < shirtCBE.CloObjectEffects.Count; i++)
-            //        {
-            //            byte partNum = (byte)shirtCBE.CloObjectEffects[i].Index;
-            //            ModelData.AddModel((byte)shirtCBE.CloObjectEffects[i].Index, (ushort)shirtCBE.CloObjectEffects[i].ModelId);
+                        for (int j = 0; j < pantsSubPal.CloSubPalettes[i].Ranges.Count; j++)
+                        {
+                            uint palOffset = pantsSubPal.CloSubPalettes[i].Ranges[j].Offset / 8;
+                            uint numColors = pantsSubPal.CloSubPalettes[i].Ranges[j].NumColors / 8;
+                            ModelData.AddPalette(pantsPal, (ushort)palOffset, (ushort)numColors);
+                        }
+                    }
+                } // end pants
 
-            //            for (int j = 0; j < shirtCBE.CloObjectEffects[i].CloTextureEffects.Count; j++)
-            //                ModelData.AddTexture((byte)shirtCBE.CloObjectEffects[i].Index, (ushort)shirtCBE.CloObjectEffects[i].CloTextureEffects[j].OldTexture, (ushort)shirtCBE.CloObjectEffects[i].CloTextureEffects[j].NewTexture);
-            //        }
+                // Get the character's initial shirt
+                uint shirtTableID = sex.ShirtList[Convert.ToInt32(character.Appearance.ShirtStyle)].ClothingTable;
+                ClothingTable shirtCT = ClothingTable.ReadFromDat(shirtTableID);
+                if (shirtCT.ClothingBaseEffects.ContainsKey(sex.SetupID))
+                {
+                    ClothingBaseEffect shirtCBE = shirtCT.ClothingBaseEffects[sex.SetupID];
+                    for (int i = 0; i < shirtCBE.CloObjectEffects.Count; i++)
+                    {
+                        byte partNum = (byte)shirtCBE.CloObjectEffects[i].Index;
+                        ModelData.AddModel((byte)shirtCBE.CloObjectEffects[i].Index, (ushort)shirtCBE.CloObjectEffects[i].ModelId);
 
-            //        // Apply the proper palette(s). Unlike character skin/hair, clothes can have several palette ranges!
+                        for (int j = 0; j < shirtCBE.CloObjectEffects[i].CloTextureEffects.Count; j++)
+                            ModelData.AddTexture((byte)shirtCBE.CloObjectEffects[i].Index, (ushort)shirtCBE.CloObjectEffects[i].CloTextureEffects[j].OldTexture, (ushort)shirtCBE.CloObjectEffects[i].CloTextureEffects[j].NewTexture);
+                    }
 
-            //        if (shirtCT.ClothingSubPalEffects.ContainsKey(character.Appearance.ShirtColor))
-            //        {
-            //            CloSubPalEffect shirtSubPal = shirtCT.ClothingSubPalEffects[character.Appearance.ShirtColor];
-            //            for (int i = 0; i < shirtSubPal.CloSubPalettes.Count; i++)
-            //            {
-            //                PaletteSet shirtPalSet = PaletteSet.ReadFromDat(shirtSubPal.CloSubPalettes[i].PaletteSet);
-            //                ushort shirtPal = (ushort)shirtPalSet.GetPaletteID(character.Appearance.ShirtHue);
+                    // Apply the proper palette(s). Unlike character skin/hair, clothes can have several palette ranges!
 
-            //                if (shirtPal > 0) // shirtPal will be 0 if the palette set is empty/not found
-            //                {
-            //                    for (int j = 0; j < shirtSubPal.CloSubPalettes[i].Ranges.Count; j++)
-            //                    {
-            //                        uint palOffset = shirtSubPal.CloSubPalettes[i].Ranges[j].Offset / 8;
-            //                        uint numColors = shirtSubPal.CloSubPalettes[i].Ranges[j].NumColors / 8;
-            //                        ModelData.AddPalette(shirtPal, (ushort)palOffset, (ushort)numColors);
-            //                    }
-            //                }
-            //            }
-            //        }
-            //    }
+                    if (shirtCT.ClothingSubPalEffects.ContainsKey(character.Appearance.ShirtColor))
+                    {
+                        CloSubPalEffect shirtSubPal = shirtCT.ClothingSubPalEffects[character.Appearance.ShirtColor];
+                        for (int i = 0; i < shirtSubPal.CloSubPalettes.Count; i++)
+                        {
+                            PaletteSet shirtPalSet = PaletteSet.ReadFromDat(shirtSubPal.CloSubPalettes[i].PaletteSet);
+                            ushort shirtPal = (ushort)shirtPalSet.GetPaletteID(character.Appearance.ShirtHue);
 
-            //    // Get the character's initial footwear
-            //    uint footwearTableID = sex.FootwearList[Convert.ToInt32(character.Appearance.FootwearStyle)].ClothingTable;
-            //    ClothingTable footwearCT = ClothingTable.ReadFromDat(footwearTableID);
-            //    if (footwearCT.ClothingBaseEffects.ContainsKey(sex.SetupID))
-            //    {
-            //        ClothingBaseEffect footwearCBE = footwearCT.ClothingBaseEffects[sex.SetupID];
-            //        for (int i = 0; i < footwearCBE.CloObjectEffects.Count; i++)
-            //        {
-            //            byte partNum = (byte)footwearCBE.CloObjectEffects[i].Index;
-            //            ModelData.AddModel((byte)footwearCBE.CloObjectEffects[i].Index, (ushort)footwearCBE.CloObjectEffects[i].ModelId);
+                            if (shirtPal > 0) // shirtPal will be 0 if the palette set is empty/not found
+                            {
+                                for (int j = 0; j < shirtSubPal.CloSubPalettes[i].Ranges.Count; j++)
+                                {
+                                    uint palOffset = shirtSubPal.CloSubPalettes[i].Ranges[j].Offset / 8;
+                                    uint numColors = shirtSubPal.CloSubPalettes[i].Ranges[j].NumColors / 8;
+                                    ModelData.AddPalette(shirtPal, (ushort)palOffset, (ushort)numColors);
+                                }
+                            }
+                        }
+                    }
+                }
 
-            //            for (int j = 0; j < footwearCBE.CloObjectEffects[i].CloTextureEffects.Count; j++)
-            //                ModelData.AddTexture((byte)footwearCBE.CloObjectEffects[i].Index, (ushort)footwearCBE.CloObjectEffects[i].CloTextureEffects[j].OldTexture, (ushort)footwearCBE.CloObjectEffects[i].CloTextureEffects[j].NewTexture);
-            //        }
+                // Get the character's initial footwear
+                uint footwearTableID = sex.FootwearList[Convert.ToInt32(character.Appearance.FootwearStyle)].ClothingTable;
+                ClothingTable footwearCT = ClothingTable.ReadFromDat(footwearTableID);
+                if (footwearCT.ClothingBaseEffects.ContainsKey(sex.SetupID))
+                {
+                    ClothingBaseEffect footwearCBE = footwearCT.ClothingBaseEffects[sex.SetupID];
+                    for (int i = 0; i < footwearCBE.CloObjectEffects.Count; i++)
+                    {
+                        byte partNum = (byte)footwearCBE.CloObjectEffects[i].Index;
+                        ModelData.AddModel((byte)footwearCBE.CloObjectEffects[i].Index, (ushort)footwearCBE.CloObjectEffects[i].ModelId);
 
-            //        // Apply the proper palette(s). Unlike character skin/hair, clothes can have several palette ranges!
-            //        CloSubPalEffect footwearSubPal = footwearCT.ClothingSubPalEffects[character.Appearance.FootwearColor];
-            //        for (int i = 0; i < footwearSubPal.CloSubPalettes.Count; i++)
-            //        {
-            //            PaletteSet footwearPalSet = PaletteSet.ReadFromDat(footwearSubPal.CloSubPalettes[i].PaletteSet);
-            //            ushort footwearPal = (ushort)footwearPalSet.GetPaletteID(character.Appearance.FootwearHue);
+                        for (int j = 0; j < footwearCBE.CloObjectEffects[i].CloTextureEffects.Count; j++)
+                            ModelData.AddTexture((byte)footwearCBE.CloObjectEffects[i].Index, (ushort)footwearCBE.CloObjectEffects[i].CloTextureEffects[j].OldTexture, (ushort)footwearCBE.CloObjectEffects[i].CloTextureEffects[j].NewTexture);
+                    }
 
-            //            for (int j = 0; j < footwearSubPal.CloSubPalettes[i].Ranges.Count; j++)
-            //            {
-            //                uint palOffset = footwearSubPal.CloSubPalettes[i].Ranges[j].Offset / 8;
-            //                uint numColors = footwearSubPal.CloSubPalettes[i].Ranges[j].NumColors / 8;
-            //                ModelData.AddPalette(footwearPal, (ushort)palOffset, (ushort)numColors);
-            //            }
-            //        }
-            //    } // end footwear
-            //}
+                    // Apply the proper palette(s). Unlike character skin/hair, clothes can have several palette ranges!
+                    CloSubPalEffect footwearSubPal = footwearCT.ClothingSubPalEffects[character.Appearance.FootwearColor];
+                    for (int i = 0; i < footwearSubPal.CloSubPalettes.Count; i++)
+                    {
+                        PaletteSet footwearPalSet = PaletteSet.ReadFromDat(footwearSubPal.CloSubPalettes[i].PaletteSet);
+                        ushort footwearPal = (ushort)footwearPalSet.GetPaletteID(character.Appearance.FootwearHue);
+
+                        for (int j = 0; j < footwearSubPal.CloSubPalettes[i].Ranges.Count; j++)
+                        {
+                            uint palOffset = footwearSubPal.CloSubPalettes[i].Ranges[j].Offset / 8;
+                            uint numColors = footwearSubPal.CloSubPalettes[i].Ranges[j].NumColors / 8;
+                            ModelData.AddPalette(footwearPal, (ushort)palOffset, (ushort)numColors);
+                        }
+                    }
+                } // end footwear 
+            } */
         }
 
         public async Task Load(AceCharacter character)
@@ -494,7 +437,6 @@ namespace ACE.Entity
             }
 
             Location = character.Location;
-
 
             IsOnline = true;
